refactor(runner): separate lifecycle state from signal handling
