+++ crates/runner/src/main.rs
@@ -22,6 +22,7 @@ mod ids;
 mod image_hash;
 mod io_limits;
 mod kmsg_log;
+mod lifecycle;
 mod live_runner_instances;
 mod local_queue;
 mod lock;
